requirements.txt
androguard==3.3.5
asn1crypto==1.4.0
attrs==20.3.0
backcall==0.2.0
biopython==1.78
bugscanner==0.1.5
cachetools==4.2.2
cattrs==1.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.22
daemonize==2.5.0
decorator==4.4.2
expertai-nlapi==2.1.3
facebook-sdk==3.1.0
Flask==2.0.1
future==0.18.2
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
kiwisolver==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
meson==0.55.3
mysql-replication==0.21
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
p==0.0.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pg-chameleon==2.0.14
pickleshare==0.7.5
Pillow==8.2.0
pproxy==2.3.5
prompt-toolkit==3.0.18
protobuf==3.17.1
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.6
pydot==1.4.2
Pygments==2.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-facebook-api==0.9.2
pytz==2021.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
responses==0.13.3
rollbar==0.15.0
rsa==4.7.2
six==1.15.0
tabulate==0.8.7
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==2.0.1
